requirements.txt
requests2==2.16.0
boto==2.45.0
panda==0.1.1
pandas==2.0.1
pandas==1.5.2
numpy==1.24.2
pandas==1.5.1